Refactor database cli
